Station-related setup + packaging

+++ rpmbuild/SPECS/hilbert-cli.spec
@@ -1,3 +1,4 @@
+####%define user          imaginary
 %global debug_package %{nil}
 %define origname      hilbert-cli
 %define user          kiosk
@@ -14,16 +15,17 @@
 # BuildRoot:      %{buildroot}
 Summary:        Hilbert: client-side tools with a basic minimal configuration
 Name:           hilbert-cli
-Version:        0.9.0
+Version:        0.9.1
 
 License:        Apache License, Version 2.0
-Release:        2%{?dist}
+Release:        7%{?dist}
 
 URL:            https://github.com/hilbert/%{origname}
-Source:         hilbert-cli.tar.gz
+Source0:         hilbert-cli.tar.gz
 # v%{version}.tar.gz
 # https://github.com/hilbert/hilbert-cli/archive/v%{version}.tar.gz
-#Source1:        OGL.tgz.tar.gz
+Source1:        hilbert-compose-customizer.tar.gz
+Source2:        docker-compose.tar.gz
 
 #hilbert-default-config.tar.gz
 #Source2:        station.cfg
@@ -45,18 +47,21 @@ Requires:       docker-engine
 
 # Group:          Development/Tools
 
+#### Moreover %{user} has to be already pre-configured before installing this package!
 %description
 Hilbert: client side tools + minimal configuration
 
 NOTE: ATM Requires internet connection to GitHub to download docker-compose 1.13.0 binary bundle. 
-Moreover %{user} has to be already pre-configured before installing this RPM!
+
 
 %prep
 ## pwd ; ls -la ; echo $RPM_BUILD_ROOT ; echo %{buildroot}; export 
 rm -rf $RPM_BUILD_ROOT
 
-%setup -c
-#?# %setup -a 1
+#%setup
+%setup -c -a 0
+%setup -D -a 1
+%setup -D -a 2
 
 ##%%%%setup -T -D -a 1
 ##%%autosetup
@@ -71,20 +76,23 @@ rm -rf $RPM_BUILD_ROOT
 
 
 # Install Hilbert Client-tools into a separate location
-mkdir -p %{buildroot}/%{_bin_dir}/
-cp tools/hilbert-station station/{docker-gc,generate_ogl.sh,get-compose.sh} %{buildroot}/%{_bin_dir}/
+mkdir -p "%{buildroot}/%{_bin_dir}/" || exit $?
+cp tools/hilbert-station station/docker-gc station/generate_ogl.sh station/get-compose.sh "%{buildroot}/%{_bin_dir}/"  || exit $?
+
 
 # Make sure that 
 # 1. hilbert-station and docker-compose can be found in the PATH
 # 2. corresponding variables: HILBERT_STATION & DOCKER_COMPOSE are set accordingly
-mkdir -p %{buildroot}/etc/profile.d/
-cp station/etc/profile.d/hilbert-env.sh %{buildroot}/etc/profile.d/
+mkdir -p "%{buildroot}/etc/profile.d/"
+cp station/etc/profile.d/hilbert-env.sh "%{buildroot}/etc/profile.d/"
 
 # Copy minimal sample configuration + sample xsession
-mkdir -p %{buildroot}/%{_cfg_dir}/
-cp -R station/station_configs/minimal %{buildroot}/%{_cfg_dir}/
-cp station/.xsession %{buildroot}/%{_cfg_dir}/
+mkdir -p "%{buildroot}/%{_cfg_dir}/"
+#cp -R station/station_configs/minimal "%{buildroot}/%{_cfg_dir}/"
+cp station/.xsession "%{buildroot}/%{_cfg_dir}/"
 
+mkdir -p "%{buildroot}/etc/tmpfiles.d/"
+echo "D /var/run/hilbert 0755 %{user} %{user} - " > "%{buildroot}/etc/tmpfiles.d/hilbert.conf"
 
 
 # OGL will be a separate story!
@@ -98,12 +106,22 @@ cp station/.xsession %{buildroot}/%{_cfg_dir}/
 #sudo -g %{user} -u %{user} HILBERT_CONFIG_BASEDIR=%{buildroot}/%{_cfg_dir}/ -i %{buildroot}/%{_bin_dir}/hilbert-station -vv init 
 #chown -R %{user}:%{user} %{buildroot}/%{_cfg_dir}/../../
 
+cp hilbert-compose-customizer "%{buildroot}/%{_bin_dir}/"  || exit $?
+
+if [ -f docker-compose ]; then
+  cp docker-compose "%{buildroot}/%{_bin_dir}/"
+fi
+
+
 # Download docker-compose 1.13.0 from GitHub 
-BASE=$PWD
-cd %{buildroot}/%{_bin_dir}/
-./get-compose.sh || exit 1 
-unlink ./compose
-cd $BASE
+if [ ! -f "%{buildroot}/%{_bin_dir}/docker-compose" ]; then
+  BASE=$PWD
+  cd %{buildroot}/%{_bin_dir}/
+  ./get-compose.sh || exit 1 
+  unlink ./compose
+  cd $BASE
+fi
+
 
 ##%%make_install
 
@@ -130,9 +148,9 @@ rm -rf $RPM_BUILD_ROOT
 # rm -f %{_cfg_dir}/configs/image
 
 ## TODO: use random temporary directory instead of /tmp/minimal or /tmp/testapp!
-sudo -g %{user} -u %{user} bash -c "cp -R %{_cfg_dir}/minimal /tmp/" && \
-sudo -g %{user} -u %{user} DOCKER_COMPOSE=%{_bin_dir}/docker-compose %{_bin_dir}/hilbert-station -q init /tmp/minimal || echo "Sorry: something failed during initialization!"
-rm -Rf /tmp/minimal
+#sudo -g %{user} -u %{user} bash -c "cp -R %{_cfg_dir}/minimal /tmp/" && \
+#sudo -g %{user} -u %{user} DOCKER_COMPOSE=%{_bin_dir}/docker-compose %{_bin_dir}/hilbert-station -q init /tmp/minimal || echo "Sorry: something failed during initialization!"
+#rm -Rf /tmp/minimal
 
 ##cd /tmp/ && (echo "0"; echo) | %{_bin_dir}/generate_ogl.sh && \
 ##sudo -g %{user} -u %{user} bash -c "cd; cp /tmp/OGL.tgz .config/hilbert-station"
@@ -147,6 +165,14 @@ rm -Rf /tmp/minimal
 # docker rmi hello-world:latest
 
 %changelog
+* Sun Oct 1 2017 Alex
+- Updated hilbert-station to avoid unnecessary output from `docker pull` 
+- Updated hilbert-compose-customizer to workaround nvidia-docker issue with the host networking driver/mode
+
+* Tue Jul 18 2017 Alex
+- Updated hilbert-station + added hilbert-compose-customizer
+
+
 * Fri May  5 2017 Alex
 - Fixed and updated hilbert-cli tools
